cura-connect/skills/hosp-automation/requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
playwright>=1.40.0
requests>=2.31.0
